% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimparser.R
\name{get_ssim_sample}
\alias{get_ssim_sample}
\title{get_ssim_sample}
\usage{
get_ssim_sample(
  datefrom = as.Date("2020-11-01"),
  dateto = as.Date("2020-12-01"),
  season = "W20"
)
}
\arguments{
\item{datefrom}{First date of the sample.}

\item{dateto}{Last date of the sample.}

\item{season}{IATA season (W20 = Winter 2020).}
}
\value{
SSIM sample as a string.
}
\description{
Get a test SSIM file for validation
}
\examples{
# Get SSIM sample
get_ssim_sample(datefrom = as.Date("2020-11-01"), dateto = as.Date("2020-12-01"), season="W20")
}
