% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ssimparser.R
\name{get_utc_time}
\alias{get_utc_time}
\title{get UTC date and time from SSIM local date time and offset}
\usage{
get_utc_time(pdate = "27NOV20", ptime = "0000", poffset = "+0000")
}
\description{
get UTC date and time from SSIM local date time and offset
}
\keyword{internal}
