% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimparser.R
\name{load_ssim}
\alias{load_ssim}
\title{load_ssim}
\usage{
load_ssim(
  ssim_file = "AUA_S_20201022_20201023_20201122.txt",
  nestres = FALSE,
  collist = get_ssim_collist(getall = FALSE),
  cleannames = TRUE,
  punpivot = TRUE,
  pexpand = FALSE
)
}
\arguments{
\item{ssim_file}{Path to the SSIM file.}

\item{nestres}{Nest SSIM type 3 into type 2 (TRUE/FALSE). Default to FALSE.}

\item{collist}{List of columns that need to be present in the final Data Frame. get_ssim_collist() to get the full list.}

\item{cleannames}{Clean column names in the final Data Frame by removing type2/type3 prefixes (TRUE/FALSE). Default TRUE.}

\item{punpivot}{Unpivot the schedules by creating a schedule by day of operation (TRUE/FALSE). Default FALSE.}

\item{pexpand}{Instantiate flights from the schedules.}
}
\value{
Data Frame (nested or not) containing the schedules (or instantiated flights).
}
\description{
Load SSIM file into a Data Frame.
}
\examples{
# Load SSIM file
ssim <- load_ssim("./AFR_20201115.txt")
}
