% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssimparser.R
\name{load_ssim_flights}
\alias{load_ssim_flights}
\title{load_ssim_flights}
\usage{
load_ssim_flights(
  ssim_files = c("AFR_20201115.txt", "AFR_20201116.txt"),
  collist = get_ssim_collist(getall = FALSE),
  cleannames = TRUE
)
}
\arguments{
\item{ssim_files}{List of SSIM files to load, in the correct order (from the first to load to the last file to load).}

\item{collist}{List of columns that need to be present in the final Data Frame. get_ssim_collist() to get the full list.}

\item{cleannames}{Clean column names in the final Data Frame by removing type2/type3 prefixes (TRUE/FALSE). Default TRUE.}
}
\value{
Data Frame containing the flights
}
\description{
Load multiple SSIM file, instantiate flights, and return the result as a Data Frame. \cr
In case of period overlap for a specific flight date, information from the latest file will be used, \cr
so beware of the file order in
parameter ssim_files.
}
\examples{
# Display the total traffic per day from two SSIM schedules files
load_ssim(c("./AFR_20201115.txt", "./AFR_20201116.txt"), cleannames = FALSE) \%>\%
group_by(flight_date = as.Date(flight.flight_date)) \%>\%
summarise(total_flights = n()) \%>\%
arrange(desc(flight_date))
}
